feat: abc for conditions
